+++ neoforge/src/main/java/com/grim3212/assorted/tools/AssortedToolsNeoForge.java
@@ -1,6 +1,7 @@
 package com.grim3212.assorted.tools;
 
 import com.grim3212.assorted.tools.client.data.ToolsLanguageProvider;
+import com.grim3212.assorted.tools.client.data.ToolsManualProvider;
 import com.grim3212.assorted.lib.data.ForgeBlockTagProvider;
 import com.grim3212.assorted.lib.data.ForgeItemTagProvider;
 import com.grim3212.assorted.lib.data.ForgeDatapackRegistryProvider;
@@ -76,5 +77,6 @@ public class AssortedToolsNeoForge {
         event.addProvider(new ToolsItemModelProvider(packOutput));
         event.addProvider(new ToolsEquipmentAssetProvider(packOutput));
         event.addProvider(new ToolsLanguageProvider(packOutput));
+        event.addProvider(new ToolsManualProvider(packOutput));
     }
 }